Fixed a bug where the Timestamped Textarea field wouldn't sort correctly.
If the timestamp was created when in minute 1-9 in any hour, an extra `0` would be added to the minutes of the timestamp value. When combined with multiple instances of the field where all instances didn't occur in minute 1-9 in any hour, the field would be sorted incorrectly and publish out of order.